+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `zig_target` option.
 - Add `zig_cpu` option.
 
+### Changed
+
+- Bump latest stable `zig` to `0.11.0`.
+
 ## [0.2.0] - 2023-07-10
 
 ### Added